Backend (Django)/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.1
django-cors-headers==4.6.0
mysqlclient==2.2.6
psycopg[binary]==3.3.3
Pillow==11.1.0
python-decouple==3.8
reportlab==4.4.1
stripe==11.6.0
